extract referrer trimming callback to a separate file
